Add new renderer using a tilesheet and a 2d rendering context.
The new renderer collects all tiles in a large tile sheet at load time and
only draws regions of the sheet when rendering of a specific tile is re-
quested.
This improves load times compared to the canvas based renderer as there's
no need to allocate and render thousands of separate image nodes. Rendering
performance however seems to suffer substantially with this approach.